requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asyncpg==0.31.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
mutagen==1.47.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
yandex-music==2.2.0
yarl==1.22.0
